Fix for noavatar.jpg gets deleted
http://www.ilch.de/forum-showposts-54453.html

+++ application/modules/user/mappers/Setting.php
@@ -47,7 +47,7 @@ class Setting extends \Ilch\Mapper
             ->execute()
             ->fetchCell('avatar');
 
-        if (file_exists($avatar)) {
+        if ((strpos($avatar, 'static/img/noavatar.jpg') == false) && file_exists($avatar)) {
             unlink($avatar);
         }
